move embeddings test to patch dir
