+++ tests/e2e/test_llama.py
@@ -5,12 +5,9 @@ E2E tests for llama
 import logging
 import os
 
-import pytest
-import torch
-
 from axolotl.cli.args import TrainerCliArgs
 from axolotl.common.datasets import load_datasets
-from axolotl.train import setup_model_and_tokenizer, train
+from axolotl.train import train
 from axolotl.utils.config import normalize_config, validate_config
 from axolotl.utils.dict import DictDefault
 
@@ -199,53 +196,3 @@ class TestLlama:
 
         train(cfg=cfg, dataset_meta=dataset_meta)
         check_model_output_exists(temp_dir, cfg)
-
-    @pytest.mark.parametrize(
-        "embeddings_skip_upcast",
-        [
-            True,
-            False,
-        ],
-    )
-    def test_peft_embeddings_upcast(self, temp_dir, embeddings_skip_upcast):
-        # pylint: disable=duplicate-code
-        cfg = DictDefault(
-            {
-                "base_model": "HuggingFaceTB/SmolLM2-135M",
-                "trust_remote_code": True,
-                "sequence_len": 512,
-                "val_set_size": 0.01,
-                "special_tokens": {
-                    "pad_token": "<|endoftext|>",
-                },
-                "datasets": [
-                    {
-                        "path": "mhenrichsen/alpaca_2k_test",
-                        "type": "alpaca",
-                    },
-                ],
-                "num_epochs": 1,
-                "max_steps": 2,
-                "micro_batch_size": 1,
-                "gradient_accumulation_steps": 1,
-                "output_dir": temp_dir,
-                "learning_rate": 0.00001,
-                "optimizer": "adamw_8bit",
-                "lr_scheduler": "cosine",
-                "flash_attention": True,
-                "sample_packing": False,
-                "bf16": "auto",
-                "save_safetensors": True,
-                "embeddings_skip_upcast": embeddings_skip_upcast,
-            }
-        )
-
-        cfg = validate_config(cfg)
-        normalize_config(cfg)
-
-        model, _, _, _ = setup_model_and_tokenizer(cfg)
-
-        # Check if the embeddings are upcast correctly
-        dtype = torch.bfloat16 if embeddings_skip_upcast else torch.float32
-        assert model.base_model.model.model.embed_tokens.weight.dtype == dtype
-        assert model.base_model.model.lm_head.weight.dtype == dtype
